Began working on a better managed file system to avoid having a main.py that keeps growing to thousands of lines long. Put the command definitions in another file and imported that file so that the main file contains only minimal function definitions
